backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
groq==0.11.0
tenacity==9.0.0
pydantic==2.9.2
python-dotenv==1.0.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
httpx==0.27.2
sse-starlette==2.1.3
pytest==8.3.3
pytest-asyncio==0.24.0
